make bigint_set_uint take uintmax_t

+++ c/mp/mp.h
@@ -48,7 +48,7 @@ int bigint_divrem(struct bigint *q, struct bigint *r, const struct bigint *a,
                   const struct bigint *b);
 
 int bigint_set(struct bigint *d, const struct bigint *s);
-int bigint_set_uint(struct bigint *a, unsigned int v);
+int bigint_set_uint(struct bigint *a, uintmax_t v);
 int bigint_set_uint1(struct bigint *a, coeff_t v);
 int bigint_mul_uint1(struct bigint *d, const struct bigint *a, coeff_t b);
